+++ src/LandingPage.test.js
@@ -1,6 +1,11 @@
 import { render } from "@testing-library/react";
 import LandingPage from "./LandingPage";
+import { MemoryRouter } from "react-router";
 
 test("renders without crashing", () => {
-  render(<LandingPage />);
+  render(
+    <MemoryRouter>
+      <LandingPage />
+    </MemoryRouter>
+  );
 });
