+++ src/Signup.test.js
@@ -1,0 +1,6 @@
+import { render } from "@testing-library/react"
+import Signup from "./Signup"
+
+test("renders without crashing", () => {
+  render(<Signup />)
+})
